LifeCycleInterface: Added onActivityCreated callback method to LifeCycleInterface.
AppLockActivity:
Changed superclass from PinActivity to Activity.
Changed code executed in onActivityResumed to onActivityCreated.
Modified finish() to only call setLastActiveMillis() if the pin REQUEST was successfull.
Call pinChallengeCancelled from onStop if the pin request was unsuccessfull.

AppLockImpl:
getSalt() was returning a longer string that the one that was saved in a previous activity. For some reason the getString method from SharedPreferences was returing some extra (not printable) charaters. Decoding and recoding the saved string fixed the problem.
Added remove(PIN_CHALLENGE_CANCELLED_PREFERENCE_KEY) to the disableAndRemoveConfiguration() method.
Changed the onActivityXXXX callbacks that did the setLastActiveMillis() lo do it only if the activity didn't currently requires a lockscreen request.